# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "description": "Generic API response.",
        "properties": {
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "description": "数据",
            "title": "Data"
          },
          "message": {
            "description": "消息",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "description": "状态",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "ApiResponse",
        "type": "object"
      },
...
      "ChatRequest": {
        "description": "Chat request payload.",
        "example": {
          "Id": "session-123",
          "Question": "什么是向量数据库？"
        },
        "properties": {
          "Id": {
            "description": "会话 ID",
            "title": "Id",
            "type": "string"
          },
          "Question": {
            "description": "用户问题",
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "Id",
          "Question"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ClearRequest": {
        "description": "Clear chat session request.",
        "properties": {
          "sessionId": {
            "description": "会话 ID",
            "title": "Sessionid",
            "type": "string"
          }
        },
        "required": [
          "sessionId"
        ],
        "title": "ClearRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionInfoResponse": {
        "description": "Chat session information response.",
        "properties": {
          "history": {
            "description": "历史消息列表",
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "History",
            "type": "array"
          },
          "message_count": {
            "description": "消息数量",
            "title": "Message Count",
            "type": "integer"
          },
          "session_id": {
            "description": "会话 ID",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "message_count",
          "history"
        ],
        "title": "SessionInfoResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI-powered SRE assistant with knowledge-grounded chat, AIOps diagnosis, and native agent workbench",
    "title": "SmartSRE Copilot",
    "version": "0.1.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "description": "快速对话接口\n{\n    \"code\": 200,\n    \"message\": \"success\",\n    \"data\": {\n        \"success\": true,\n        \"answer\": \"回答内容\",\n        \"errorMessage\": null\n    }\n}\n\nArgs:\n    request: 对话请求\n\nReturns:\n    统一格式的对话响应",
        "operationId": "Chat-chat-post-api_v1_chat",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/clear": {
      "post": {
        "description": "清空会话历史\n\nArgs:\n    request: 清空请求\n\nReturns:\n    操作结果",
        "operationId": "Chat-clear_session-post-api_v1_chat_clear",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClearRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Session",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/session/{session_id}": {
      "get": {
        "description": "查询会话历史\n\nArgs:\n    session_id: 会话 ID\n\nReturns:\n    会话信息",
        "operationId": "Chat-get_session_info-get-api_v1_chat_session_session_id",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session Info",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/session/{session_id}/tool-events": {
      "get": {
        "description": "查询会话工具调用事件。",
        "operationId": "Chat-get_session_tool_events-get-api_v1_chat_session_session_id_tool_events",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session Tool Events",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/sessions": {
      "get": {
        "description": "列出已持久化的会话摘要。",
        "operationId": "Chat-list_sessions-get-api_v1_chat_sessions",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/stream": {
      "post": {
        "description": "流式对话接口（基于 RAG Agent，SSE）\n\n返回 SSE 格式，data 字段为 JSON：\n\n工具调用事件:\nevent: message\ndata: {\"type\":\"tool_call\",\"data\":{\"tool\":\"工具名\",\"status\":\"start|end\",\"input\":{...}}}\n\n内容流式事件:\nevent: message\ndata: {\"type\":\"content\",\"data\":\"内容块\"}\n\n完成事件:\nevent: message\ndata: {\"type\":\"done\",\"data\":{\"answer\":\"完整答案\",\"tool_calls\":[...]}}\n\nArgs:\n    request: 对话请求\n\nReturns:\n    SSE 事件流",
        "operationId": "Chat-chat_stream-post-api_v1_chat_stream",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Stream",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/contracts/openapi": {
      "get": {
        "operationId": "Contracts-get_openapi_contract-get-api_v1_contracts_openapi",
        "parameters": [
          {
            "in": "query",
            "name": "include_spec",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Spec",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Openapi Contract",
        "tags": [
          "Contracts"
        ]
      }
    },
...
  }
}